requirements.txt
absl-py==1.4.0
altgraph==0.17.3
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
decorator==5.1.1
facenet-pytorch==2.5.2
Flask==2.2.3
Flask-Cors==3.0.10
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.11.0
Keras-Applications==1.0.8
keras-vggface==0.6
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
mtcnn==0.1.1
npyscreen==4.10.5
numpy==1.24.2
oauthlib==3.2.2
opencv-python-headless==4.5.5.62
opt-einsum==3.3.0
packaging==23.0
pefile==2023.2.7
Pillow==9.4.0
protobuf==3.19.6
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.0
pyOpenSSL==23.1.0
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
python-version==0.0.2
pyuac==0.0.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
tee==0.0.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.5.0